+++ packages/brick-types/src/manifest.ts
@@ -693,6 +693,8 @@ export interface BaseEntityResolveConf {
 
   /** {@inheritDoc HandleReject} */
   onReject?: HandleReject;
+
+  /** {@inheritDoc BuiltinBrickEventHandler.if} */
   if?: string | boolean;
 }
 
@@ -726,7 +728,7 @@ export type RefResolveConf = Pick<
 };
 
 /** {@inheritDoc HandleRejectByTransform} */
-export type HandleReject = HandleRejectByTransform /*| HandleRejectByCatch*/;
+export type HandleReject = HandleRejectByTransform | HandleRejectByCatch;
 
 /**
  * 当异步数据处理抛出错误时，使用该错误对象作为数据源进行转换来作为最终结果。
@@ -736,9 +738,13 @@ export interface HandleRejectByTransform {
   transform: GeneralTransform;
 }
 
-// export interface HandleRejectByCatch {
-//   catch: true;
-// }
+/**
+ * 当异步数据处理抛出错误时，捕获它，可以设置只对应的构件报错，而不是页面整体报错。
+ */
+export interface HandleRejectByCatch {
+  /** 只让构件报错。 */
+  isolatedCrash: true;
+}
 
 /**
  * 数据转换配置。
@@ -1346,22 +1352,37 @@ export type CustomTemplateProxyProperty =
   | CustomTemplateProxyTransformableProperty;
 
 /** 自定义模板的基本属性代理配置。 */
-export interface CustomTemplateProxyBasicProperty {
+export interface CustomTemplateProxyBasicProperty
+  extends CustomTemplateProxyWithExtra {
   /** 要代理的构件的引用 ID。 */
   ref: string;
 
-  /** 要代理的构件的属性名 */
+  /** 要代理的构件的属性名。 */
   refProperty: string;
 }
 
 /** 自定义模板的可转换的属性代理配置。 */
-export interface CustomTemplateProxyTransformableProperty {
+export interface CustomTemplateProxyTransformableProperty
+  extends CustomTemplateProxyWithExtra {
   /** {@inheritDoc CustomTemplateProxyBasicProperty.ref} */
   ref: string;
 
   /** 要代理的构件的属性转换设置。 */
   refTransform: GeneralTransform;
 }
+
+export interface CustomTemplateProxyWithExtra {
+  /** {@inheritDoc CustomTemplateProxyExtraOneWayRef} */
+  extraOneWayRefs?: CustomTemplateProxyExtraOneWayRef[];
+}
+
+/** 额外的单向（由模板传输给内部构件）属性代理配置。 */
+export type CustomTemplateProxyExtraOneWayRef =
+  | Omit<CustomTemplateProxyBasicProperty, keyof CustomTemplateProxyWithExtra>
+  | Omit<
+      CustomTemplateProxyTransformableProperty,
+      keyof CustomTemplateProxyWithExtra
+    >;
 
 /**
  * 自定义模板的事件代理配置表。