+++ dll/d3/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.4](https://git.easyops.local/anyclouds/next-core/compare/@dll/d3@2.0.3...@dll/d3@2.0.4) (2020-11-02)
+
+**Note:** Version bump only for package @dll/d3
+
+## [2.0.3](https://git.easyops.local/anyclouds/next-core/compare/@dll/d3@2.0.2...@dll/d3@2.0.3) (2020-10-29)
+
+**Note:** Version bump only for package @dll/d3
+
 ## [2.0.2](https://git.easyops.local/anyclouds/next-core/compare/@dll/d3@2.0.1...@dll/d3@2.0.2) (2020-10-28)
 
 **Note:** Version bump only for package @dll/d3