+++ packages/brick-dll/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.4](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.3.3...@easyops/brick-dll@2.3.4) (2020-11-03)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [2.3.3](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.3.2...@easyops/brick-dll@2.3.3) (2020-11-03)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [2.3.2](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.3.1...@easyops/brick-dll@2.3.2) (2020-11-02)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
+## [2.3.1](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.3.0...@easyops/brick-dll@2.3.1) (2020-10-29)
+
+**Note:** Version bump only for package @easyops/brick-dll
+
 # [2.3.0](https://git.easyops.local/anyclouds/next-core/compare/@easyops/brick-dll@2.2.7...@easyops/brick-dll@2.3.0) (2020-10-28)
 
 ### Features